filetree: make actions placed in somewhat saner locations
the three actions belonging to the filetree have been removed from the menubar.

- the up/down actions are keyboard-centric in nature, since a user with a pointing device
  has the actual file list to click at, so they don't really need to be shown in the mouse-centric
  part of the UI.
- the 'show active' has been moved to the file list, since 'go > show active' away from the file list
  doesn't really make it clear what exactly it's going to show active in. this location is marginally
  more convenient for mouse users, but really, this is about making it more legible.

the save and save as actions have been removed from the minitoolbar there to make room for the new show active
action. they already have the most recognisable keyboard shortcut in the history of keyboard shortcuts, a presence
in the menubar, main toolbar, and in the command palette. i'm sure they can do without another incarnation in the file tree.