KateFileTree: tweak close button handling
Only show the close-button column in the headerview when the button is set
to be shown in the config.

Set the minimum section size to 16 (SizeSmall), so that ResizeMode::ResizeToContents
strictly takes up only the contents size.

Remove check for header() pointer, QTreeView has a header by default, and
other parts of the code don't check for header().