Handle enabling/disabling of search action in SearchPlugin
Usecase:

I should be able to use the Ctrl+Shift+F shortcut in other places e.g.,
Konsole plugin. Atm the shortcut in SearchPlugin is global which
prevents this.

The shortcut does need to be global to be useful but it doesn't have to
be active always. In fact, we can disable it if KTE::View doesn't have
focus.

Previous solution disabled it in the project plugin which only makes it
work in one place.